Auto-reconnect the last connected wallet (#101)
Each time a wallet is connected, we save the last connected wallet ID in the local storage. When loading the app page, we first retrieve that latest wallet ID, then we discover available wallets. If we find that the last connected wallet is present in discovered wallets, and is already "enabled" (no need to approve anymore a wallet connection) then we auto-reconnect the wallet.

This is a QoL improvement, saving few clicks for users.